requirements.txt
streamlit==1.32.2
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
supabase==2.3.4
PyMuPDF==1.23.22
google-generativeai==0.3.2
uuid